JIT: Enable phi-based jump threading when SSA updates aren't needed (#77748)
Leverage the new SSA accounting to look for cases of phi-based jump threading
that will not require SSA updates. In particular cases where the phis are all
locally consumed.

Also update documentation on the SSA checker implementation (which aims to
ensure that the SSA accounting we're relying on here is accurate).