Remove "wprintf" from PAL (#77852)
The last usage was in FormatMessageW, which
only used wprintf syntax for messages that used
embedded printf format between two "!". We do
not use this feature in any of our messages so
this code path is technically unused.

Remove _snwprintf_s from PAL

Remove _vsnwprintf_s from PAL

_woutput_s was not removed because it is intertwined
with the implementation for printf. The linker will just
prune the implementation anyways.

FormatMessageW errors if embedded formatting is used
Update PAL tests for FormatMessageW