Standardize test script naming to CLRTestBa{tc,s}hP{re,ost}Commands (#76483)
Main change is BashCLRTestPreCommands -> CLRTestBashPreCommands. As part of this I noticed that wasm tests overwrite the Bash precommands, and fixing that breaks them so it isn't touched here. Also, there are several non-Commands variables that start with Ba{tc,s}hTest that aren't touched here.

Remove check on RunCrossGen2 in disasm checks as that is intended as a shell environment variable, not an msbuild variable.

Convert a few more tests to the new environment variable format.