Sync linker to NativeAOT (#77894)
Sync latest changes from dotnet/linker to the shared code in NativeAOT and also to the code which is similar in NativeAOT.

Syncs NativeAOT to https://github.com/dotnet/linker/tree/e502e7255030dde029e984e397924338485d1079/

The only notable change is the switch to use ParameterProxy and related abstractions when operating with method parameters. This was introduced in linker to avoid bugs when using a simple integer to refer to parameters by index, since different parts of the system use different indexing (some have index 0 as the implicit this parameter, while others don't).

The places which are similar to code in linker I converted to use ParameterProxy and helpers. Other places which interact directly with the type system in NativeAOT codebase I kept using the underlying APIs as it seems to make more sense.

One test change - this modifies the infra for running linker tests to define `NATIVEAOT` when compiling any of the tests sources. This is needed to ifdef out a test targeting vararg which is not supported by NativeAOT.