Reduce host memory usage: don't keep resolved deps around for app lifetime (#77987)
`hostpolicy` keeps a global `hostpolicy_init_t` around which has a list of `fx_definition_t` which each contain an associated `deps_json_t`. The deps are resolved/parsed on initialization of hostpolicy, updating the globally-stored information with all the resolved assets and computing properties that will be passed to the runtime.

The only part of the deps resolution results we currently use after initialization / computing the properties is the rid fallback graph of the root framework (for component dependency resolution). Everything else we are just holding on to for the lifetime of the application. 

This separates `deps_json_t` from `fx_definition_t` such that we:
- don't keep all the resolved deps around for the lifetime of the application
- store only the one rid fallback graph we need for future resolutions
- stop including/building deps parsing functionality in `hostfxr` (via `hostcommon`)

For a blank console app (so no other dependencies, just NETCoreApp) running on Windows, this reduces the amount we keep on the native heap by ~200kB. The more dependencies, the more we were keeping around.