Add feature attributes design (dotnet/linker#3103)
* Add feature attributes design

This is an attempt to specify the behavior of the various
`Requires`* attributes we have today. This is intended to specify
the allowed and disallowed semantics for a generalized feature
attribute, not the specific formatting of the warnings.

Commit migrated from https://github.com/dotnet/linker/commit/99dfd625930330003a8ae248b1d95ed3a5d5d51c