Fix GetCustomAttributeByName generic attributes (#79103)
The current logic doesn't handle the case where
a TypeSpec can be a generic attribute. Since C#
now supports generic attributes, this is a deficiency
that can be observed by diagnostics tooling using
IMetaDataImport.

The TypeSpec specification is defined in II.23.2.14 of
ECMA-335. Practically, CustomAttributes only support
TypeDef, TypeRef and TypeSpec with GENERICINST.
This means the check could be more targeted if desired.