Generic warning location tests (dotnet/linker#3146)
Adds a test which focused on warnings produced due to annotated generic parameters. This is relatively tricky for the product to deal with since it's often part of the type system and not code. So this test focused on all the different situations such warning can occur and the location where the warning is generated.

Co-authored-by: Tlakaelel Axayakatl Ceja <tlakaelel.ceja@microsoft.com>

Commit migrated from https://github.com/dotnet/linker/commit/da244b780a1d1755cb7cb389df599fe650726892