Use tokens for validation of Kept members in linker tests for NativeAOT (#79228)
Switches from using display names (which are currently not consitent between linker/AOT) to tokens. This is possible for NativeAOT because the analysis reports type system entities as read from the input assembly, and thus we can directly compare them to the input assembly read via Cecil.

(Note that this is not possible in linker version of this infra, because the linked members are read from the output assembly which has different token numbers).

This also enables validation of properties and events.