[wasm] Add the illink substitutions for SIMD (#79672)
This should fix size regression introduce with https://github.com/dotnet/runtime/pull/78068

The untrimmed S.N.Vector class added few kilobytes to the assemblies
of the bench sample. That wasn't that bad, OTOH the code produced
by the aot compiler was much larger, 5MB in this case.

So this will case default non-SIMD case. In the case of SIMD build we
will need more investigation.