Enable regex generator nullable reference types validation (#80142)
* Enable regex generator nullable reference types

Multiple times in the past we've audited the regex source generator for nullable reference type annotations, but as it was only compiled officially for netstandard2.0, any work we did to get it green eventually rotted.  This follows the work done to enable also building a .NET Core target for the json generator to do the same for the regex generator.  In doing so, I also took the opportunity to clean up the REGEXGENERATOR compilation constant.

* Add SetTargetFramework to ProjectReferences

* Fix regex unit tests