[mono][interp] Fix issue with clearing of unused defs (#80186)
If within a basic block we have code like `somevar = val1; ...; somevar = val2` and somevar has no indirects and is not used between the two defines, then we can remove the first assignment. However, MINT_SRC_DST_OFF is a special opcode that writes only to a part of a valuetype, even though it has the vt var as a destination. This means we can't clear the previous define, otherwise we lose data in part of the valuetype.