Remove remaining use of handles in interp_exec_method and instead use a reserved slot at the top of the stack (#80197)
Right now every time we enter the interpreter we waste some CPU time allocating a handle, even though the only remaining use of the handle is in the BOX opcodes. This PR instead allocates some space at the top of the stack for a temporary GC-visible object pointer, and then uses that instead of a handle.