Add more precise tracking of static bases (#80093)
Under some circumstances, the compiler needs to generate information about the bases where static fields of a type are placed.

The existing logic was simple: if a static base was accessed, generate the owning type. If owning type was generated, generate all of it's static bases. At object emission, go over all generated types and preserve information about their bases.

We couldn't do better at the time this was implemented.

The information about static bases is needed in two cases:

1. A static field on the type is reflectable. We now know when that happens and it doesn't happen very often.
2. The generic type whose static base we generated has a template and we could build the type at runtime. This also doesn't happen very often anymore.

This PR teaches the compiler to only emit the information for these two cases. It prevents generating unnecessary EETypes, static bases, and .cctors.

Also adding a test case for scenario 2 above since we didn't have coverage for the scenario. Scenario 1 is covered.

Saves 0.2% on a hello world. Not much, but at least it's easier to reason about why this is generated.