Add more specialized frozen collection types. (#79794)
* Add more specialized frozen collection types.

- Add SmallFrozenDictionary/Set which don't do any hashing and simply iterate through arrays,
testing each element

- Add SparseRangeInt32FrozenSet which uses a bit vector for storage.

FrozenDictionary/Set each have some constants defined to drive the heuristics that
decide when to use the small and sparse collection types. I did some preliminary tuning
of these, but this could use a detailed benchmark to pick the right values. I'll leave
the fine-tuning to a separate PR.

* Expand supported numeric sets on .NET 7+

Co-authored-by: Martin Taillefer <mataille@microsoft.com>