Remove `CustomAttributeData` usage in IsByRefLike (#80841)
Grabbing custom attribute data brings the entire reflection stack (we need properties, constructors, fields, methods).

We ask `IsByRefLike` in type loader constraint validation. Ideally we should be able to get rid of the type loader from a hello world, but that's for later. This is a more efficient implementation either way.