Add an optimized C implementation of STFLD_O that uses the correct kind of write barrier (#81806)
This PR moves most of the jiterpreter's STFLD_O implementation into a C function that is responsible for also doing the null check. As a bonus that function is able to use the correct kind of write barrier (though it's not clear to me whether the previous one was broken in any way).