Fixing a FATAL_GC_ERROR in loh_allocate_in_condemned when in extremely low memory situations (#82418)
I'm running GCPerfSim with this commandline and this env -

corerun.exe GCPerfSim.dll -tc 60 -tagb 100 -tlgb 9.5 -lohar 0 -pohar 0 -sohsr 100-4000 -lohsr 102400-204800 -pohsr 100-204800 -sohsi 50 -lohsi 0 -pohsi 0 -sohpi 0 -lohpi 0  -sohfi 0 -lohfi 0 -pohfi 0 -allocType simple -testKind time

set COMPlus_GCServer=1
set COMPlus_GCHeapCount=1e
set COMPlus_GCHeapHardLimit=280000000

this causes us to hit this FATAL_GC_ERROR. We have one object which is at the very end of the region and we are asking for its size and padding of 2 * loh_padding_obj_size. since (heap_segment_allocated - heap_segment_mem) is (size + loh_padding_obj_size), loh_size_fit_p reports its not fitting so we call grow_heap_segment and at this point we are so low on memory we don't have ~16 pages left so we try to goto the next region and there's no more regions (in this case this is the only region in gen3).

We shouldn't be asking for (2 * loh_padding_obj_size) for padding because this is at the end of a region (same with segments).